♻️ extract git.utils.js
