+++ scripts/lib/git.utils.js
@@ -1,0 +1,24 @@
+const os = require('os')
+const fs = require('fs')
+
+const { command, getSecretKey } = require('./utils')
+
+function initGitConfig(repository) {
+  const githubDeployKey = getSecretKey('ci.browser-sdk.github_deploy_key')
+  const homedir = os.homedir()
+
+  // ssh-add expects a new line at the end of the PEM-formatted private key
+  // https://stackoverflow.com/a/59595773
+  command`ssh-add -`.withInput(`${githubDeployKey}\n`).run()
+  command`mkdir -p ${homedir}/.ssh`.run()
+  command`chmod 700 ${homedir}/.ssh`.run()
+  const sshHost = command`ssh-keyscan -H github.com`.run()
+  fs.appendFileSync(`${homedir}/.ssh/known_hosts`, sshHost)
+  command`git config user.email ci.browser-sdk@datadoghq.com`.run()
+  command`git config user.name ci.browser-sdk`.run()
+  command`git remote set-url origin ${repository}`.run()
+}
+
+module.exports = {
+  initGitConfig,
+}
