+++ scripts/lib/utils.js
@@ -1,5 +1,3 @@
-const os = require('os')
-const fs = require('fs')
 const childProcess = require('child_process')
 const spawn = require('child_process').spawn
 // node-fetch v3.x only support ESM syntax.
@@ -12,22 +10,6 @@ function getSecretKey(name) {
   `
     .run()
     .trim()
-}
-
-function initGitConfig(repository) {
-  const githubDeployKey = getSecretKey('ci.browser-sdk.github_deploy_key')
-  const homedir = os.homedir()
-
-  // ssh-add expects a new line at the end of the PEM-formatted private key
-  // https://stackoverflow.com/a/59595773
-  command`ssh-add -`.withInput(`${githubDeployKey}\n`).run()
-  command`mkdir -p ${homedir}/.ssh`.run()
-  command`chmod 700 ${homedir}/.ssh`.run()
-  const sshHost = command`ssh-keyscan -H github.com`.run()
-  fs.appendFileSync(`${homedir}/.ssh/known_hosts`, sshHost)
-  command`git config user.email ci.browser-sdk@datadoghq.com`.run()
-  command`git config user.name ci.browser-sdk`.run()
-  command`git remote set-url origin ${repository}`.run()
 }
 
 /**
@@ -210,7 +192,6 @@ async function fetchWrapper(url, options) {
 
 module.exports = {
   getSecretKey,
-  initGitConfig,
   command,
   spawnCommand,
   printError,
