+++ scripts/staging-ci/staging-reset.js
@@ -1,8 +1,9 @@
 'use strict'
 
 const fs = require('fs')
-const { initGitConfig, command, printLog, runMain } = require('../lib/utils')
+const { command, printLog, runMain } = require('../lib/utils')
 const { CI_FILE, replaceCiFileVariable, readCiFileVariable } = require('../lib/files.utils')
+const { initGitConfig } = require('../lib/git.utils')
 
 const REPOSITORY = process.env.GIT_REPOSITORY
 const MAIN_BRANCH = process.env.MAIN_BRANCH
