+++ scripts/test/bump-chrome-driver-version.js
@@ -1,8 +1,9 @@
 'use strict'
 
 const fs = require('fs')
-const { printLog, printError, runMain, command, initGitConfig, getSecretKey, fetch } = require('../lib/utils')
+const { printLog, printError, runMain, command, getSecretKey, fetch } = require('../lib/utils')
 const { CI_FILE, replaceCiFileVariable } = require('../lib/files.utils')
+const { initGitConfig } = require('../lib/git.utils')
 
 const REPOSITORY = process.env.GIT_REPOSITORY
 const MAIN_BRANCH = process.env.MAIN_BRANCH
